requirements.txt
nats-py==2.4.0
